Refactored ScoreHolder and LabelType to their own files
